make pages dynamic
